requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.0.3
django_coverage_plugin==1.2.2
Sphinx
django-extensions==1.6.1
Werkzeug==0.11.4
django-test-plus==1.0.12
factory_boy==2.6.1
django_fakery==1.3.1
unicodecsv==0.14.1

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.4

# improved REPL
ipdb==0.9.0

# pytest!
pytest-cov==2.2.1
pytest-django==2.9.1
pytest-sugar==0.5.1

# linters
pydocstyle==1.0.0
